MessageHandlers moved to their package
