+++ src/main/java/ru/hh/resumebuilderbot/Answer.java
@@ -1,19 +1,19 @@
 package ru.hh.resumebuilderbot;
 
-class Answer {
+public class Answer {
     private ChatId chatId;
     private Object answerBody;
 
-    Answer(ChatId chatId, Object answerBody) {
+    public Answer(ChatId chatId, Object answerBody) {
         this.chatId = chatId;
         this.answerBody = answerBody;
     }
 
-    ChatId getChatId() {
+    public ChatId getChatId() {
         return chatId;
     }
 
-    Object getAnswerBody() {
+    public Object getAnswerBody() {
         return answerBody;
     }
 }
