--- src/main/java/ru/hh/resumebuilderbot/MessageHandler.java
@@ -1,5 +1,0 @@
-package ru.hh.resumebuilderbot;
-
-public interface MessageHandler {
-    public QuestionGenerator handle(Answer answer);
-}
