+++ src/main/java/ru/hh/resumebuilderbot/MessageHandler/MessageHandler.java
@@ -1,0 +1,8 @@
+package ru.hh.resumebuilderbot.MessageHandler;
+
+import ru.hh.resumebuilderbot.Answer;
+import ru.hh.resumebuilderbot.QuestionGenerator;
+
+public interface MessageHandler {
+    public QuestionGenerator handle(Answer answer);
+}
