--- src/main/java/ru/hh/resumebuilderbot/ShowMessageHandler.java
@@ -1,8 +1,0 @@
-package ru.hh.resumebuilderbot;
-
-public class ShowMessageHandler implements MessageHandler {
-    @Override
-    public QuestionGenerator handle(Answer answer) {
-        return new ShowAllQuestionGenerator();
-    }
-}
