--- src/main/java/ru/hh/resumebuilderbot/UnknownMessageHandler.java
@@ -1,9 +1,0 @@
-package ru.hh.resumebuilderbot;
-
-
-public class UnknownMessageHandler implements MessageHandler {
-    @Override
-    public QuestionGenerator handle(Answer answer) {
-        return new FixedQuestionGenerator(TextsStorage.getText("Unknown"));
-    }
-}
